requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.4
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
langgraph==0.0.20
faiss-cpu==1.7.4
google-generativeai==0.8.3